moved preprocessing data into its own folder - if not passed a date, it filters to last months data
