+++ analyzefire.py
@@ -159,86 +159,6 @@ def analyzeFire(fireDF):
     Dataframe:
         A larger dataframe with additional information gleamed from the passed file.
     """
-
-    # =================================================================
-    #     Basic Format
-    # =================================================================
-    # correct naming conventions
-    # ------------------
-    renames = {
-        "Master_Incident_Number": "Master Incident Number",
-        "Last Unit Clear Incident": "Last Real Unit Clear Incident",
-        "X_Long": "X-Long",
-    }
-    fireDF = fireDF.rename(columns=renames, errors="ignore")
-
-    # replace all instances of "-" with null values
-    # ------------------
-    fireDF = fireDF.replace("-", np.nan)
-
-    # convert strings to datetime
-    # ------------------
-    time_columns_to_convert = [
-        "Earliest Time Phone Pickup AFD or EMS",
-        "Incident Time Call Entered in Queue",
-        "Time First Real Unit Assigned",
-        "Time First Real Unit Enroute",
-        "Incident Time First Staged",
-        "Time First Real Unit Arrived",
-        "Incident Time Call Closed",
-        "Last Real Unit Clear Incident",
-    ]
-
-    for index, colName in enumerate(time_columns_to_convert):
-        pd.to_datetime(fireDF[colName], errors="raise", unit="s")
-        # group and move to front for testing
-        # fireDF = utils.putColAt(fireDF, [colName], index)
-
-    # convert strings/dates to timeDelta
-    # ------------------
-
-    timeDeltasToConvert = [
-        "Earliest Time Phone Pickup to 1st Real Unit Assigned",
-        "Incident Turnout - 1st Real Unit Assigned to 1st Real Unit Enroute",
-        "Incident First Unit Response - 1st Real Unit Assigned to 1st Real Unit Arrived",
-        "Earliest Time Phone Pickup to 1st Real Unit Arrived",
-        "Time Spent OnScene - 1st Real Unit Arrived to Last Real Unit Call Cleared",
-        "Unit Dispatch to Respond Time",
-        "Unit Respond to Arrival",
-        "Unit Dispatch to Onscene",
-        "Unit OnScene to Clear Call",
-        "Unit Assign To Clear Call Time",
-    ]
-
-    def tryDelta(times):
-        if type(times) in [float, int, np.float64]:
-            return times
-        return pd.Timedelta(str(times)) / np.timedelta64(1, "s")
-
-    for index, colName in enumerate(timeDeltasToConvert):
-        # fireDF[colName] = fireDF.apply(lambda x: tryDelta(x[colName]), axis=1)
-        fireDF[colName] = np.vectorize(tryDelta)(fireDF[colName])
-        # group and move to front for testing
-        fireDF = utils.putColAt(fireDF, [colName], index)
-
-    #
-
-    # now that we have actual times... trim the file to a reasonable size
-    print("triming dates")
-    start = pd.to_datetime("10/01/2021")
-    end = pd.to_datetime("11/01/2021")
-    fireDF = fireDF[
-        (fireDF["Earliest Time Phone Pickup AFD or EMS"] >= start)
-        & (fireDF["Earliest Time Phone Pickup AFD or EMS"] < end)
-        | (fireDF["Incident Time Call Entered in Queue"] >= start)
-        & (fireDF["Incident Time Call Entered in Queue"] < end)
-    ]
-
-    # Reset index for this subset
-    fireDF = fireDF.reset_index(drop=True)
-
-    show(fireDF)
-
     # =================================================================
     #     Match esri formatting for first arrived
     # =================================================================
@@ -263,25 +183,6 @@ def analyzeFire(fireDF):
     from GUI.validateData import checkFile
 
     fireDF = checkFile(fireDF)
-
-    # =================================================================
-    #     order fire data by time
-    # =================================================================
-    # - 'Master Incident Number' > 'Unit Time Arrived At Scene' > 'Unit Time Staged' > 'Unit Time Enroute' > 'Unit Time Assigned'
-    fireDF = fireDF.sort_values(
-        by=[
-            "Master Incident Number",
-            "Unit Time Arrived At Scene",
-            "Unit Time Staged",
-            "Unit Time Enroute",
-            "Unit Time Assigned",
-        ]
-    )
-    fireDF = fireDF.reset_index(drop=True)
-    # =================================================================
-    #     Drop useless data
-    # =================================================================
-    # fireDF = fireDF.drop(["ESD02_Record"])
 
     # =================================================================
     #     get Complete Response Force for each Structure Fire
